Refactor: 친구 리스트에서 좌측 친구 정보를 컴포넌트화
